Add empty-state report coverage
